@@ -12,7 +12,7 @@
         "fast-xml-parser": "^3.19.0",
         "he": "^1.2.0",
         "https": "^1.0.0",
-        "next": ">=10.1.2",
+        "next": "^10.2.0",
         "prettier": "^2.2.1",
         "react": "17.0.2",
         "react-dom": "17.0.2",
